Add channel to signify server readiness
Added an optional channel param which can be used to signify that the
server is ready to accept connection. Run function will close the
channel (if supplied) when it is ready to accept connection.

User of this function can choose to wait for signal from Run function by
supplying a channel of struct type, and waiting for it to return a
signal. This is especially useful for tests to avoid timing the startup
of server.